1. Push max block size to 64 MB (minus 4 bytes). 2. Remove 'stream closed' check before writing bits. Disallow writing by resetting fields in the close() code so that any attempt to write will trigger an exception/error.  Add test to check that reading from / writing to a closed stream fails. 3. Remove Sync() from OutputStram interface in Go.  Add a utility class to create buffered streams from files in Go and use it in the Block (de)compressor classes. Delete output file generated during speed test in the test class.  Tweak bitstream java code. 4. Rename ZLT to (a more proper) ZRLT: zero run length transform.

+++ go/src/kanzi/io/InfoPrinter.go
@@ -115,7 +115,7 @@ func (this *InfoPrinter) ProcessEvent(evt *BlockEvent) {
 		// Add percentage for encoding
 		if this.type_ == ENCODING {
 			if bi.stage0Size != 0 {
-				msg += fmt.Sprintf(" (%d%%)", stage2Size*100/bi.stage0Size)
+				msg += fmt.Sprintf(" (%d%%)", uint64(stage2Size)*100/uint64(bi.stage0Size))
 			}
 		}
